refactor(liveobjects): move WireObjectData.toCompactJsonElement to its type's file
The toCompactJsonElement() extension only reads WireObjectData's own fields and is
used across packages (path, livemap), so it belongs with the type it extends rather
than in the livemap feature file. Move it into message/WireObjectMessage.kt alongside
the other WireObjectData extensions (size, isInvalid); retarget the DefaultPathObject
import and drop the now-unused gson imports from InternalLiveMap. Pure move, no
behaviour change; message still has no dependency on value/path/livemap.
