+++ liveobjects/src/main/kotlin/io/ably/lib/liveobjects/value/livemap/InternalLiveMap.kt
@@ -11,10 +11,7 @@ import io.ably.lib.liveobjects.value.BaseRealtimeObject
 import io.ably.lib.liveobjects.value.ObjectType
 import io.ably.lib.liveobjects.value.ObjectUpdate
 import io.ably.lib.liveobjects.value.livecounter.DefaultLiveCounter
-import com.google.gson.JsonElement
-import com.google.gson.JsonNull
 import com.google.gson.JsonObject
-import com.google.gson.JsonPrimitive
 import java.util.concurrent.ConcurrentHashMap
 import java.util.AbstractMap
 
@@ -235,17 +232,4 @@ internal class InternalLiveMap private constructor(
       return InternalLiveMap(objectId, objects)
     }
   }
-}
-
-/**
- * Compact-JSON form of a primitive wire leaf: primitives become JsonPrimitives, binary
- * stays base64-encoded (RTPO14b1), embedded JSON passes through.
- */
-internal fun WireObjectData.toCompactJsonElement(): JsonElement = when {
-  string != null -> JsonPrimitive(string)
-  number != null -> JsonPrimitive(number)
-  boolean != null -> JsonPrimitive(boolean)
-  bytes != null -> JsonPrimitive(bytes) // base64 kept encoded - RTPO14b1
-  json != null -> json!!
-  else -> JsonNull.INSTANCE // unreachable for a valid Leaf (isInvalid-filtered)
 }
